Brew install: Add test for package.list

+++ scripts/10_brew-cask/install.sh
@@ -8,14 +8,16 @@ if [ "$OS" != "osx" ]; then exit; fi
 brew install brew-cask
 
 # Install any package given in `packages.list` (one per line). See: http://stackoverflow.com/questions/10929453/bash-scripting-read-file-line-by-line
-while read -r package
-do
-  echo "Installing $package"
-  brew cask install "$package"
-done < "packages.list"
+if [ -f "packages.list" ]; then
+  while read -r package
+  do
+    echo "Installing $package"
+    brew cask install "$package"
+  done < "packages.list"
+fi
 
 # Install ALL software packages if 'init.sh --full' is called.
-if [ "$FULL" = true ]; then
+if [ "$FULL" = true -a -f "packages.full.list" ]; then
   while read -r package
   do
     echo "Installing $package"
